Split the PostGraphile config out of the PostGraphile middleware
