+++ @app/server/src/plugins/PassportLoginPlugin.ts
@@ -1,6 +1,6 @@
 import { gql, makeExtendSchemaPlugin } from "graphile-utils";
 
-import { OurGraphQLContext } from "../middleware/installPostGraphile";
+import { OurGraphQLContext } from "../graphile.config";
 import { ERROR_MESSAGE_OVERRIDES } from "../utils/handleErrors";
 
 const PassportLoginPlugin = makeExtendSchemaPlugin((build) => ({
